refactor: 패스 구조 변경

+++ src/main/java/wooteco/subway/auth/ui/AuthenticationPrincipalArgumentResolver.java
@@ -3,9 +3,6 @@ package wooteco.subway.auth.ui;
 import java.util.Map;
 import java.util.Objects;
 import javax.servlet.http.HttpServletRequest;
-
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
 import org.springframework.core.MethodParameter;
 import org.springframework.web.bind.support.WebDataBinderFactory;
 import org.springframework.web.context.request.NativeWebRequest;
@@ -32,16 +29,16 @@ public class AuthenticationPrincipalArgumentResolver implements HandlerMethodArg
 
     @Override
     public MemberRequest resolveArgument(MethodParameter parameter,
-                                        ModelAndViewContainer mavContainer,
-                                        NativeWebRequest webRequest,
-                                        WebDataBinderFactory binderFactory) {
+        ModelAndViewContainer mavContainer,
+        NativeWebRequest webRequest,
+        WebDataBinderFactory binderFactory) {
         HttpServletRequest httpServletRequest =
             webRequest.getNativeRequest(HttpServletRequest.class);
         String payLoad =
             AuthorizationExtractor.extract(Objects.requireNonNull(httpServletRequest));
         Map<String, Object> data = authService.getPayLoadData(payLoad);
 
-        Long id = (long) (int) data.get("id");
+        Long id = (long) data.get("id");
         String email = (String) data.get("email");
 
         return new MemberRequest(id, email);
